+++ imo_vmdb/csv_import/magnitudes.py
@@ -40,10 +40,10 @@ class MagnitudesParser(CsvParser):
 
     def __init__(self, *args, **kwars):
         super().__init__(*args, **kwars)
-        self._delete_stmt = self._db_conn.convert_stmt(
+        self._delete_stmt = self._db_conn._convert_stmt(
             "DELETE FROM imported_magnitude WHERE id = %(id)s"
         )
-        self._insert_stmt = self._db_conn.convert_stmt("""
+        self._insert_stmt = self._db_conn._convert_stmt("""
             INSERT INTO imported_magnitude (
                 id,
                 observer_id,
@@ -67,7 +67,7 @@ class MagnitudesParser(CsvParser):
         if self._do_delete:
             try:
                 cur.execute(
-                    self._db_conn.convert_stmt("DELETE FROM imported_magnitude")
+                    self._db_conn._convert_stmt("DELETE FROM imported_magnitude")
                 )
             except Exception as e:
                 raise DBException(str(e))