+++ imo_vmdb/csv_import/session.py
@@ -23,10 +23,10 @@ class SessionParser(CsvParser):
 
     def __init__(self, *args, **kwars):
         super().__init__(*args, **kwars)
-        self._delete_stmt = self._db_conn.convert_stmt(
+        self._delete_stmt = self._db_conn._convert_stmt(
             "DELETE FROM imported_session WHERE id = %(id)s"
         )
-        self._insert_stmt = self._db_conn.convert_stmt("""
+        self._insert_stmt = self._db_conn._convert_stmt("""
             INSERT INTO imported_session (
                 id,
                 observer_id,
@@ -51,7 +51,7 @@ class SessionParser(CsvParser):
     def on_start(self, cur):
         if self._do_delete:
             try:
-                cur.execute(self._db_conn.convert_stmt("DELETE FROM imported_session"))
+                cur.execute(self._db_conn._convert_stmt("DELETE FROM imported_session"))
             except Exception as e:
                 raise DBException(str(e))